move logging functionalities to the `io.logger` module.
