+++ src/dfastbe/utils.py
@@ -1,50 +1,6 @@
 from typing import List, Tuple
-import time
 import numpy as np
 
-def timed_logger(label: str) -> None:
-    """
-    Write a message with time information.
-
-    Arguments
-    ---------
-    label : str
-        Message string.
-    """
-    time, diff = _timer()
-    print(time + diff + label)
-
-
-def _timer() -> Tuple[str, str]:
-    """
-    Return text string representation of time since previous call.
-
-    The routine uses the global variable LAST_TIME to store the time of the
-    previous call.
-
-    Arguments
-    ---------
-    None
-
-    Returns
-    -------
-    time_str : str
-        String representing duration since first call.
-    diff_str : str
-        String representing duration since previous call.
-    """
-    global FIRST_TIME
-    global LAST_TIME
-    new_time = time.time()
-    if "LAST_TIME" in globals():
-        time_str = "{:6.2f} ".format(new_time - FIRST_TIME)
-        diff_str = "{:6.2f} ".format(new_time - LAST_TIME)
-    else:
-        time_str = "   0.00"
-        diff_str = "       "
-        FIRST_TIME = new_time
-    LAST_TIME = new_time
-    return time_str, diff_str
 
 def get_zoom_extends(
         km_min: float,
